Refactor code structure and improve functionality
- Renamed files and classes for consistency (e.g., questionDone.dart to question_done.dart).
- Updated imports to reflect new naming conventions.
- Enhanced error handling in AuthService by using rethrow for exceptions.
- Simplified constructors in various widgets by using `super.key`.
- Added PlayerProvider to manage player-related logic and state.
- Introduced AnsweredQuestionsProvider to track answered questions.
- Implemented QuestionBoardPage to display questions and manage game state.
- Improved UI components for better user experience and error handling.
- Added logging for better debugging and tracking of application state.
