+++ lib/config/secrets.dart
@@ -1,4 +1,4 @@
 class Secrets {
-  static const GSheetAPI =
+  static const gSheetAPI =
       "https://script.google.com/macros/s/AKfycbxi9pnH_cNPPAh56c-Y8VHJZDzadSXcSOw5cfq5ukTmn8KhLhgG2fUit5jogaCQC1iC6A/exec";
 }
