+++ lib/widgets/app_background.dart
@@ -4,7 +4,7 @@ import 'dart:ui';
 class AppBackground extends StatelessWidget {
   final Widget child;
 
-  const AppBackground({Key? key, required this.child}) : super(key: key);
+  const AppBackground({super.key, required this.child});
 
   @override
   Widget build(BuildContext context) {
