+++ web/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
 
 <head>
   <!--
@@ -15,11 +15,9 @@
     This is a placeholder for base href that will be replaced by the value of
     the `--base-href` argument provided to `flutter build`.
   -->
-  <base href="$FLUTTER_BASE_HREF">
-
   <meta charset="UTF-8">
+  <base href="$FLUTTER_BASE_HREF">
   <meta name="viewport" content="width=device-width,initial-scale=1.0">
-  <meta content="IE=Edge" http-equiv="X-UA-Compatible">
   <meta name="description" content="A new Flutter project.">
 
   <!-- iOS meta tags & icons -->
@@ -32,7 +30,7 @@
   <link rel="icon" type="image/png" href="favicon.png" />
 
   <title>buzz5_quiz_app</title>
-  <link rel="manifest" href="manifest.json">
+  <link rel="manifest" href="manifest.webmanifest">
 
   <script type="module">
     // Import the functions you need from the SDKs you need
